+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Please refer to :doc:`release notes<release_notes>`.
 - Enable optional tablib dependencies (#1647)
 - Removed unused method ``utils.original()``
 - Clarified ``skip_diff`` documentation (#1655)
+- Standardised interface of :meth:`~import_export.widgets.Widget.render` (#1657)
 - Improved documentation relating to validation on import (#1665)
 
 4.0.0-alpha.5 (2023-09-22)